+++ src/main/java/br/com/tecchallenge/resources/SubjectResource.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import br.com.tecchallenge.exceptions.ResourceNotFoundException;
 import br.com.tecchallenge.models.Subject;
 import br.com.tecchallenge.services.SubjectService;
 
@@ -34,6 +35,9 @@ public class SubjectResource {
 	
 	@GetMapping("/subject/{id}")
 	public Subject findSubject(@PathVariable(value="id") long id){
+		if (subjectService.findSubject(id) == null)
+			throw new ResourceNotFoundException("Pauta " + id + " não encontrada.");
+		
 		return subjectService.findSubject(id);
 	}
 	
@@ -44,16 +48,25 @@ public class SubjectResource {
 	
 	@DeleteMapping("/subject")
 	public void deleteSubject(@RequestBody @Valid Subject subject) {
+		if (subjectService.findSubject(subject.getId()) == null)
+			throw new ResourceNotFoundException("Pauta " + subject.getId() + " não encontrada.");
+		
 		subjectService.deleteSubject(subject);
 	}
 	
 	@PutMapping("/subject")
 	public Subject updateSubject(@RequestBody @Valid Subject subject) {
+		if (subjectService.findSubject(subject.getId()) == null)
+			throw new ResourceNotFoundException("Pauta " + subject.getId() + " não encontrada.");
+		
 		return subjectService.updateSubject(subject);
 	}
 	 
 	@GetMapping("/subject/{id}/result")
 	public Subject findSubjectResult(@PathVariable(value="id") long id){
+		if (subjectService.findSubject(id) == null)
+			throw new ResourceNotFoundException("Pauta " + id + " não encontrada.");
+		
 		return subjectService.findSubjectResult(id);
 	}
 }
